DAOS-2660: move pool API to daos_pool.h
Move pool types and API to a new header file daos_pool.h
