+++ src/include/daos.h
@@ -29,7 +29,7 @@
 #include <daos_array.h>
 #include <daos_kv.h>
 #include <daos_prop.h>
-
+#include <daos_pool.h>
 #include <daos_mgmt.h>
 #include <daos_security.h>
