+++ src/include/daos_types.h
@@ -145,170 +145,6 @@ daos_handle_is_inval(daos_handle_t hdl)
 }
 
 /**
- * Storage Targets
- */
-
-/** Type of storage target */
-typedef enum {
-	DAOS_TP_UNKNOWN,
-	/** Rotating disk */
-	DAOS_TP_HDD,
-	/** Flash-based */
-	DAOS_TP_SSD,
-	/** Persistent memory */
-	DAOS_TP_PM,
-	/** Volatile memory */
-	DAOS_TP_VM,
-} daos_target_type_t;
-
-/** Current state of the storage target */
-typedef enum {
-	DAOS_TS_UNKNOWN,
-	/* Not available */
-	DAOS_TS_DOWN_OUT,
-	/* Not available, may need rebuild */
-	DAOS_TS_DOWN,
-	/* Up */
-	DAOS_TS_UP,
-	/* Up and running */
-	DAOS_TS_UP_IN,
-} daos_target_state_t;
-
-/** Description of target performance */
-typedef struct {
-	/** TODO: storage/network bandwidth, latency etc */
-	int			foo;
-} daos_target_perf_t;
-
-enum {
-	DAOS_MEDIA_SCM	= 0,
-	DAOS_MEDIA_NVME,
-	DAOS_MEDIA_MAX
-};
-
-/** Pool target space usage information */
-struct daos_space {
-	/* Total space in bytes */
-	uint64_t	s_total[DAOS_MEDIA_MAX];
-	/* Free space in bytes */
-	uint64_t	s_free[DAOS_MEDIA_MAX];
-};
-
-/** Target information */
-typedef struct {
-	daos_target_type_t	ta_type;
-	daos_target_state_t	ta_state;
-	daos_target_perf_t	ta_perf;
-	struct daos_space	ta_space;
-} daos_target_info_t;
-
-/** Pool space usage information */
-struct daos_pool_space {
-	/* Aggregated space for all live targets */
-	struct daos_space	ps_space;
-	/* Min target free space in bytes */
-	uint64_t	ps_free_min[DAOS_MEDIA_MAX];
-	/* Max target free space in bytes */
-	uint64_t	ps_free_max[DAOS_MEDIA_MAX];
-	/* Average target free space in bytes */
-	uint64_t	ps_free_mean[DAOS_MEDIA_MAX];
-	/* Target(VOS) count */
-	uint32_t	ps_ntargets;
-	uint32_t	ps_padding;
-};
-
-struct daos_rebuild_status {
-	/** pool map version in rebuilding or last completed rebuild */
-	uint32_t		rs_version;
-	/** Time (Seconds) for the rebuild */
-	uint32_t		rs_seconds;
-	/** errno for rebuild failure */
-	int32_t			rs_errno;
-	/**
-	 * rebuild is done or not, it is valid only if @rs_version is non-zero
-	 */
-	int32_t			rs_done;
-
-	/* padding of rebuild status */
-	int32_t			rs_padding32;
-
-	/* Failure on which rank */
-	int32_t			rs_fail_rank;
-	/** # total to-be-rebuilt objects, it's non-zero and increase when
-	 * rebuilding in progress, when rs_done is 1 it will not change anymore
-	 * and should equal to rs_obj_nr. With both rs_toberb_obj_nr and
-	 * rs_obj_nr the user can know the progress of the rebuilding.
-	 */
-	uint64_t		rs_toberb_obj_nr;
-	/** # rebuilt objects, it's non-zero only if rs_done is 1 */
-	uint64_t		rs_obj_nr;
-	/** # rebuilt records, it's non-zero only if rs_done is 1 */
-	uint64_t		rs_rec_nr;
-
-	/** rebuild space cost */
-	uint64_t		rs_size;
-};
-
-/**
- * Pool info query bits.
- * The basic pool info like fields from pi_uuid to pi_leader will always be
- * queried for each daos_pool_query() calling. But the pi_space and
- * pi_rebuild_st are optional based on pi_mask's value.
- */
-enum daos_pool_info_bit {
-	/** true to query pool space usage */
-	DPI_SPACE		= 1ULL << 0,
-	/** true to query rebuild status */
-	DPI_REBUILD_STATUS	= 1ULL << 1,
-	/** query all above optional info */
-	DPI_ALL			= -1,
-};
-
-/**
- * Storage pool
- */
-typedef struct {
-	/** Pool UUID */
-	uuid_t				pi_uuid;
-	/** Number of targets */
-	uint32_t			pi_ntargets;
-	/** Number of nodes */
-	uint32_t			pi_nnodes;
-	/** Number of deactivated targets */
-	uint32_t			pi_ndisabled;
-	/** Latest pool map version */
-	uint32_t			pi_map_ver;
-	/** current raft leader */
-	uint32_t			pi_leader;
-	/** pool info bits, see daos_pool_info_bit */
-	uint64_t			pi_bits;
-	/** Space usage */
-	struct daos_pool_space		pi_space;
-	/** rebuild status */
-	struct daos_rebuild_status	pi_rebuild_st;
-} daos_pool_info_t;
-
-/*
- * DAOS management pool information
- */
-typedef struct {
-	/* TODO? same pool info structure as a pool query?
-	 * requires back-end RPC to each pool service.
-	 * daos_pool_info_t		 mgpi_info;
-	 */
-	uuid_t				 mgpi_uuid;
-	/** List of current pool service replica ranks */
-	d_rank_list_t			*mgpi_svc;
-} daos_mgmt_pool_info_t;
-
-/* DAOS pool container information
- *
- */
-struct daos_pool_cont_info {
-	uuid_t		pci_uuid;
-};
-
-/**
  * DAOS_PC_RO connects to the pool for reading only.
  *
  * DAOS_PC_RW connects to the pool for reading and writing.
@@ -329,7 +165,6 @@ struct daos_pool_cont_info {
 /**
  * Epoch
  */
-
 typedef uint64_t	daos_epoch_t;
 
 typedef struct {
